requirements.txt
SQLAlchemy>=2.0
psycopg2-binary>=2.9
pydantic>=1.10
pytest>=7.0
pytest-asyncio>=0.21.0
alembic>=1.11    
python-dotenv>=1.0.0